Corrections

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,18 +8,18 @@
  */
 size_t print_list(const list_t *h)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	while (h != NULL)
 	{
 		if (h->str == NULL)
 		{
 
-			printf("[%u] %s\n", 0, "(nil)");
+			printf("[%d] %s\n", 0, "(nil)");
 		}
 		else
 		{
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 		}
 		h = h->next;
 		i++;
